ai_ta_backend/rabbitmq/requirements.txt
psycopg2
openai<2.0.0
pandas
tiktoken>=0.7
boto3>=1.28.79
qdrant-client>=1.9.1
langchain>=0.0.331
posthog>=3.1.0
pysrt>=1.1.2
docx2txt>=0.8
pydub>=0.25.1
ffmpeg-python>=0.2.0
ffprobe>=0.5
ffmpeg>=1.4
PyMuPDF>=1.23.6
pytesseract>=0.3.10
openpyxl>=3.1.2
networkx>=3.2.1
python-pptx>=0.6.23
unstructured>=0.18.9
langchain-community
msoffcrypto-tool>=5.2.0
GitPython>=3.1.40
beautifulsoup4>=4.12.2
sentry-sdk>=1.39.1
pdfplumber>=0.11.0
pika>=1.3.1
aio-pika>=9.0.5
python-dotenv>=1.1.0
Flask-SQLAlchemy>=3.1.1
canvasapi>=3.2.0
bs4>=0.0.2
yt-dlp>=2024.7.16
azure-storage-blob
langchain_openai>=0.3.5
langchain-core
langchain-qdrant
langchain-text-splitters
